+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
-- set_component to Entity
+- 'set' (component) to Entity
 - property 'entities' to System
 - System constructor raises TypeError if a callable is not passed
 - pool property to Entity